hpc_scripts/requirements_lora.txt
torch
torchvision
diffusers==0.27.2
transformers
accelerate
datasets
safetensors
pillow
peft
tqdm
